+++ packages/module/tests/uniqueTypes.test.ts
@@ -1,4 +1,4 @@
-import { uniqueTypes } from '../src/module'
+import { uniqueTypes } from '../src/utils'
 import { test } from '@fast-check/vitest'
 import { arbitrary, types } from '@mondrian-framework/model'
 import { describe, expect } from 'vitest'
